Move defined functions and variables to utils.py
